Add some text to test VS code commit messages
We have changed our default editor for git commit messages to VScode.
This should make it al ot easier to write commit messages, especially
multiline messages with head and body. We might also make some mistakes
to edit later. Let's see if that works.

Automatic wrapping is something we need to figure out though, since this
seems to be a bit weird